# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trip {
  id: number;
  name: string;
  destination: string;
  description: string;
  priceCzk: number;
  days: number;
  /** @nullable */
  imageUrl?: string | null;
  availableSpots: number;
  active: boolean;
  depositAmount?: number;
}

export interface TripInput {
  name: string;
  destination: string;
  description: string;
  priceCzk: number;
  days: number;
  imageUrl?: string;
  availableSpots?: number;
  active?: boolean;
}

export interface TripUpdate {
  name?: string;
  destination?: string;
  description?: string;
  priceCzk?: number;
  days?: number;
  imageUrl?: string;
  availableSpots?: number;
  active?: boolean;
}

export type BookingBookingType =
  (typeof BookingBookingType)[keyof typeof BookingBookingType];

export const BookingBookingType = {
  deposit: "deposit",
  full: "full",
} as const;

export type BookingPaymentMethod =
  (typeof BookingPaymentMethod)[keyof typeof BookingPaymentMethod];

export const BookingPaymentMethod = {
  card: "card",
  qr: "qr",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  deposit_paid: "deposit_paid",
  fully_paid: "fully_paid",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  tripId: number;
  trip?: Trip;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  bookingType: BookingBookingType;
  paymentMethod: BookingPaymentMethod;
  amountCzk: number;
  status: BookingStatus;
  /** @nullable */
  stripeSessionId?: string | null;
  createdAt: string;
}

export type BookingInputBookingType =
  (typeof BookingInputBookingType)[keyof typeof BookingInputBookingType];

export const BookingInputBookingType = {
  deposit: "deposit",
  full: "full",
} as const;

export type BookingInputPaymentMethod =
  (typeof BookingInputPaymentMethod)[keyof typeof BookingInputPaymentMethod];

export const BookingInputPaymentMethod = {
  card: "card",
  qr: "qr",
} as const;

export interface BookingInput {
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  bookingType: BookingInputBookingType;
  paymentMethod: BookingInputPaymentMethod;
}

export type BookingStatusUpdateStatus =
  (typeof BookingStatusUpdateStatus)[keyof typeof BookingStatusUpdateStatus];

export const BookingStatusUpdateStatus = {
  pending: "pending",
  deposit_paid: "deposit_paid",
  fully_paid: "fully_paid",
  cancelled: "cancelled",
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface StripeSessionResult {
  url: string;
}

export interface QrPaymentDetails {
  bookingId: number;
  iban: string;
  recipientName: string;
  amountCzk: number;
  variableSymbol: string;
  qrBase64: string;
  tripName: string;
  customerEmail: string;
}

export interface AdminStats {
  totalBookings: number;
  pendingBookings: number;
  paidBookings: number;
  totalRevenueCzk: number;
  activeTrips: number;
}