# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalBookings: number;
  pendingBookings: number;
  paidBookings: number;
  totalRevenueCzk: number;
  activeTrips: number;
}