# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingBookingType } from "./bookingBookingType";
import type { BookingPaymentMethod } from "./bookingPaymentMethod";
import type { BookingStatus } from "./bookingStatus";
import type { Trip } from "./trip";

export interface Booking {
  id: number;
  tripId: number;
  trip?: Trip;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerPhone?: string | null;
  bookingType: BookingBookingType;
  paymentMethod: BookingPaymentMethod;
  amountCzk: number;
  status: BookingStatus;
  /** @nullable */
  stripeSessionId?: string | null;
  createdAt: string;
}