# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputBookingType } from "./bookingInputBookingType";
import type { BookingInputPaymentMethod } from "./bookingInputPaymentMethod";

export interface BookingInput {
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  bookingType: BookingInputBookingType;
  paymentMethod: BookingInputPaymentMethod;
}