# lib/api-zod/src/generated/types/qrPaymentDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */

export interface QrPaymentDetails {
  bookingId: number;
  iban: string;
  recipientName: string;
  amountCzk: number;
  variableSymbol: string;
  qrBase64: string;
  tripName: string;
  customerEmail: string;
}