# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */

export interface Trip {
  id: number;
  name: string;
  destination: string;
  description: string;
  priceCzk: number;
  days: number;
  /** @nullable */
  imageUrl?: string | null;
  availableSpots: number;
  active: boolean;
  depositAmount?: number;
}