# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kyrgyzstán Zájezdy API
 * OpenAPI spec version: 0.1.0
 */

export interface TripUpdate {
  name?: string;
  destination?: string;
  description?: string;
  priceCzk?: number;
  days?: number;
  imageUrl?: string;
  availableSpots?: number;
  active?: boolean;
}